chore: test chars in RFC3986 not covered by encodeURIComponent

+++ src/node.js
@@ -13,10 +13,11 @@ const { REGION } = require("./config");
   const params = {
     Bucket: bucketName
   };
-  const objectNames = [
-    "NameWithoutParentheses.jpg",
-    "NameWith(Parentheses).jpg"
-  ];
+  const prefix = "test-";
+  const suffix = ".jpg";
+
+  // Refs: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encodeURIComponent#Description
+  const rfc3986ExtendedChars = "!'()*";
 
   const v2Client = new AWS.S3({ region: REGION });
   const v3Client = new S3({ region: REGION });
@@ -28,13 +29,18 @@ const { REGION } = require("./config");
   console.log(`\nWaiting for "${bucketName}" bucket creation\n`);
   await v2Client.waitFor("bucketExists", params).promise();
 
-  for (let i = 0; i < objectNames.length; i++) {
-    console.log(`Putting "${objectNames[i]}" in ${bucketName}`);
-    await v3Client.putObject({
-      Body: "000000",
-      Bucket: bucketName,
-      Key: objectNames[i]
-    });
+  for (let i = 0; i < rfc3986ExtendedChars.length; i++) {
+    const Key = `${prefix}${rfc3986ExtendedChars.charAt(i)}${suffix}`;
+    console.log(`Putting "${Key}" in ${bucketName}`);
+    try {
+      await v3Client.putObject({
+        Body: "000000",
+        Bucket: bucketName,
+        Key
+      });
+    } catch (e) {
+      console.log(`PutObject "${Key}" failed...`);
+    }
   }
 
   response = await v3Client.listObjects(params);
@@ -42,12 +48,15 @@ const { REGION } = require("./config");
   console.log(JSON.stringify(response, null, 2));
   console.log("\n");
 
-  for (let i = 0; i < objectNames.length; i++) {
-    console.log(`Deleting "${objectNames[i]}" from ${bucketName}`);
-    await v3Client.deleteObject({
-      Bucket: bucketName,
-      Key: objectNames[i]
-    });
+  if (response.Contents) {
+    for (let i = 0; i < response.Contents.length; i++) {
+      const { Key } = response.Contents[i];
+      console.log(`Deleting "${Key}" from ${bucketName}`);
+      await v3Client.deleteObject({
+        Bucket: bucketName,
+        Key
+      });
+    }
   }
 
   console.log(`\nDeleting bucket ${bucketName}`);
